+++ packages/apidom-ls/src/config/asyncapi/message-trait/lint/allowed-fields-3-0.ts
@@ -1,9 +1,35 @@
+import { DiagnosticSeverity } from 'vscode-languageserver-types';
+
+import ApilintCodes from '../../../codes.ts';
 import { LinterMeta } from '../../../../apidom-language-types.ts';
+import { AsyncAPI3 } from '../../target-specs.ts';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 const allowedFields3_0Lint: LinterMeta = {
-  // add remaining document
-  targetSpecs: [{ namespace: 'asyncapi', version: '3.0.0' }],
+  code: ApilintCodes.NOT_ALLOWED_FIELDS,
+  source: 'apilint',
+  message: 'Object includes not allowed fields',
+  severity: DiagnosticSeverity.Error,
+  linterFunction: 'allowedFields',
+  linterParams: [
+    [
+      'headers',
+      'correlationId',
+      'contentType',
+      'name',
+      'title',
+      'summary',
+      'description',
+      'tags',
+      'externalDocs',
+      'bindings',
+      'examples',
+      '$ref',
+    ],
+    'x-',
+  ],
+  marker: 'key',
+  targetSpecs: AsyncAPI3,
 };
 
 export default allowedFields3_0Lint;