+++ packages/apidom-ls/src/config/asyncapi/operation/lint/index.ts
@@ -3,6 +3,8 @@ import operationIdUniqueLint from './operation-id--unique.ts';
 import summaryTypeLint from './summary--type.ts';
 import descriptionTypeLint from './description--type.ts';
 import securityItemsType2_0__2_6Lint from './security--items-type-2-0--2-6.ts';
+import securityType3_0Lint from './security--type-3-0.ts';
+import securityItemsType3_0Lint from './security--items-type-3-0.ts';
 import tagsTypeLint from './tags--type.ts';
 import externalDocsType from './external-docs--type.ts';
 import bindingsTypeLint from './bindings--type.ts';
@@ -13,6 +15,13 @@ import messagesTypeLint from './messages--type.ts';
 import allowedFields2_0__2_3Lint from './allowed-fields-2-0--2-3.ts';
 import allowedFields2_4__2_6Lint from './allowed-fields-2-4--2-6.ts';
 import allowedFields3_0Lint from './allowed-fields-3-0.ts';
+import actionTypeLint from './action--type.ts';
+import actionRequiredLint from './action--required.ts';
+import channelTypeLint from './channel--type.ts';
+import channelRequiredLint from './channel--required.ts';
+import titleTypeLint from './title--type.ts';
+import replyTypeLint from './reply--type.ts';
+import $refValidLint from './$ref--valid.ts';
 
 const lints = [
   operationIdTypeLint,
@@ -20,6 +29,8 @@ const lints = [
   summaryTypeLint,
   descriptionTypeLint,
   securityItemsType2_0__2_6Lint,
+  securityType3_0Lint,
+  securityItemsType3_0Lint,
   tagsTypeLint,
   externalDocsType,
   bindingsTypeLint,
@@ -30,6 +41,13 @@ const lints = [
   allowedFields2_0__2_3Lint,
   allowedFields2_4__2_6Lint,
   allowedFields3_0Lint,
+  actionTypeLint,
+  actionRequiredLint,
+  channelTypeLint,
+  channelRequiredLint,
+  titleTypeLint,
+  replyTypeLint,
+  $refValidLint,
 ];
 
 export default lints;