+++ packages/apidom-ls/src/config/asyncapi/parameter/lint/enum--type.ts
@@ -1,0 +1,20 @@
+import { DiagnosticSeverity } from 'vscode-languageserver-types';
+
+import ApilintCodes from '../../../codes.ts';
+import { LinterMeta } from '../../../../apidom-language-types.ts';
+import { AsyncAPI3 } from '../../target-specs.ts';
+
+const enumTypeLint: LinterMeta = {
+  code: ApilintCodes.ASYNCAPI3_PARAMETER_FIELD_ENUM_TYPE,
+  source: 'apilint',
+  message: "'enum' must be an array of strings",
+  severity: DiagnosticSeverity.Error,
+  linterFunction: 'apilintArrayOfType',
+  linterParams: ['string'],
+  marker: 'key',
+  target: 'enum',
+  data: {},
+  targetSpecs: AsyncAPI3,
+};
+
+export default enumTypeLint;